turns out lots of things e.g. cv2 don't like np.int64 for a dtype for a grayscale image, which i can understand. change to np.uint8

+++ model_ed2.py
@@ -80,7 +80,7 @@ imgs_all = []
 for idx, r in train_only_all_points.iterrows():
     imgs_all.append(
         np.array(r["Image"].split())
-        .astype(np.int64)
+        .astype(np.uint8)
         .reshape(IMAGE_WIDTH, IMAGE_HEIGHT, 1)
     )
 imgs_all = np.array(imgs_all)
